+++ examples/ExampleClientAccountUsers.php
@@ -26,11 +26,11 @@
  * PHP Version 5.3
  *
  * @category  Tune
- * 
+ *
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.12
+ * @version   $Date: 2014-11-05 14:59:02 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  */
@@ -68,6 +68,10 @@ class ExampleClientAccountUsers
             throw new \InvalidArgumentException("Parameter 'api_key' is not defined.");
         }
 
+        echo "=========================================================" . PHP_EOL;
+        echo "= Tune Management API Client Account Users              =" . PHP_EOL;
+        echo "=========================================================" . PHP_EOL;
+
         $client = new \Tune\Management\Shared\Service\TuneManagementClient(
             $controller = 'account/users',
             $action = 'find.json',
@@ -80,7 +84,13 @@ class ExampleClientAccountUsers
 
         $client->call();
 
-        echo $client->getResponse()->toString() . PHP_EOL;
+        echo "= TuneManagementResponse:" . PHP_EOL;
+        echo print_r($client->getResponse(), true) . PHP_EOL;
+
+        echo "======================================================" . PHP_EOL;
+        echo " End Example                                          " . PHP_EOL;
+        echo "======================================================" . PHP_EOL;
+        echo PHP_EOL;
     }
 }