+++ examples/TuneExamplesAutoloader.php
@@ -30,11 +30,12 @@
  * PHP Version 5.3
  *
  * @category  Tune
- * 
+ *
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
+ * @package   tune_examples_autoloader
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.12
+ * @version   $Date: 2014-11-03 15:03:06 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  * Autoloader for Tune MobileAppTracking Management API files.
@@ -42,11 +43,8 @@
  */
 
 
-
 /**
  * Tune SDK Examples Autoloader Class
- *
- * @package Tune\Examples
  */
 class TuneExamplesAutoloader
 {