+++ src/Tune/Management/Api/Account.php
@@ -25,11 +25,12 @@
  * PHP Version 5.3
  *
  * @category  Tune
- * 
+ *
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
+ * @package   management_api_account
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.12
+ * @version   $Date: 2014-11-03 15:03:06 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  */
@@ -40,8 +41,6 @@ use Tune\Management\Shared\Endpoints\ItemsEndpointBase;
 
 /**
  * Tune Management API endpoint '/account/'
- *
- * @package Tune\Management\Api
  */
 class Account extends ItemsEndpointBase
 {