+++ src/Tune/Management/Api/Advertiser/Stats.php
@@ -26,11 +26,12 @@
  * PHP Version 5.3
  *
  * @category  Tune
- * 
+ *
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
+ * @package   management_api_advertiser_stats
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.12
+ * @version   $Date: 2014-11-03 15:03:06 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  */
@@ -41,8 +42,6 @@ use Tune\Management\Shared\Endpoints\ReportsActualsEndpointBase;
 
 /**
  * Tune Management API endpoint '/advertiser/stats/'
- *
- * @package Tune\Management\Api\Advertiser
  *
  * @example ExampleReportsActuals.php
  */