+++ src/Tune/Management/Api/Export.php
@@ -26,11 +26,12 @@
  * PHP Version 5.3
  *
  * @category  Tune
- * 
+ *
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
+ * @package   management_api_export
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.12
+ * @version   $Date: 2014-11-03 15:03:06 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  */
@@ -45,8 +46,6 @@ use Tune\Shared\TuneServiceException;
 /**
  * Provides status of report export request, and upon completion provides
  * download url.
- *
- * @package Tune\Management\Api
  */
 class Export extends EndpointBase
 {