+++ src/Tune/Management/Shared/Service/TuneManagementRequest.php
@@ -26,11 +26,12 @@
  * PHP Version 5.3
  *
  * @category  Tune
- * 
+ *
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
+ * @package   management_shared_service_request
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.12
+ * @version   $Date: 2014-11-03 15:03:06 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  */
@@ -81,7 +82,7 @@ class TuneManagementRequest
 
     /**
      * Constructor
-     * 
+     *
      * @param string $controller        Tune Management API controller
      * @param string $action            Tune Management API controller's action
      * @param string $api_key           User's API Key provide by their MobileAppTracking platform account.