+++ src/TuneApi.php
@@ -29,11 +29,11 @@
  * PHP Version 5.3
  *
  * @category  Tune
- * 
+ *
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.12
+ * @version   $Date: $today ${today} 2014-11-03 14:00:00 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  */